Show RGB of System Color. RIght-click Popup.
Call GetSysColor() to get the RGB color and show in Status line, buton tip, and clipboard text. Handy if you like the color and want to use it or adjust it.

Right click on button Popup allows setting sample text or background to color. "..." buttons next to Sample allow picking with ColorDialog().

Example of read Only ENTRY with Font color as Gray Text so looks disabled but can still selec and copy text.